Add resiliency to slurm job reporting gaps (#118)
Method
Request the job info from FireCREST up to 5 times to add resiliency against gaps in SLURM reporting.

Rationale
After asking around I now understand that SLURM itself can have (very short) gaps of time around state changes (at least from PENDING to RUNNING) where sacct does not report on the job (this is what FireCREST uses internally). This is not specific to FireCREST, however it may be more rare to observe with other forms of transport due to latency differences. As these time gaps are usually below one second, my solution does not add a wait time between retries (network latency should be enough).

Possible future improvements
Technically in the case where we are requesting info for a specific job, which was only just submitted, we should be able to retry until we get a non-empty response (or an error). However this would introduce additional complexity for ensuring that it is a recently submitted job (and not a very old one, in case the daemon has not been started for years), as I don't think there are any guarantees from sacct for how long old jobids will be reported on.

I would only attempt that in case this solution turns out to be not enough.

---------
Co-authored-by: Ali <khsrali@gmail.com>